moved list to App.tsx and created interface
